+++ auth_rest/lib/auth_rest.dart
@@ -22,7 +22,11 @@ export 'package:tekartik_firebase_auth_rest/src/auth_rest_token.dart'
         RestAuthTokenRefreshException,
         idTokenExpiration;
 export 'package:tekartik_firebase_auth_rest/src/auth_service_rest.dart'
-    show firebaseAuthServiceRest, FirebaseAuthServiceRest, authServiceRest;
+    show
+        firebaseAuthServiceRest,
+        firebaseAuthServiceRestAdmin,
+        FirebaseAuthServiceRest,
+        authServiceRest;
 export 'package:tekartik_firebase_auth_rest/src/google_auth_rest.dart'
     show GoogleAuthOptions, GoogleRestAuthProvider;
 export 'package:tekartik_firebase_persistence/firebase_persistence.dart'